fix: show data protection notice on publish
