+++ features/dashboard/components/Stories/PublishButton.vue
@@ -5,6 +5,8 @@ import { ref, computed } from 'vue';
 
 import { backendUrl } from '../../../../store/contantsDataNarrator.js';
 
+import ConfirmPublishPopup from './ConfirmPublishPopup.vue';
+
 const { t } = useTranslation();
 
 const props = defineProps({
@@ -17,6 +19,7 @@ const emit = defineEmits([
 ]);
 
 const loading = ref(false);
+const confirmPublishOpen = ref(false);
 
 const isDraft = computed(() => props.isDraft);
 const icon = computed(() => (isDraft.value ? mdiCloudUploadOutline : mdiCloudOffOutline));
@@ -26,7 +29,7 @@ const btnLabel = computed(() =>
     : t('additional:modules.dataNarrator.button.unpublish', 'Unpublish')
 );
 
-async function onClick() {
+async function publish() {
   if (loading.value) return;
   loading.value = true;
   try {
@@ -41,11 +44,24 @@ async function onClick() {
     loading.value = false;
   }
 }
+
+function publishOk() {
+  confirmPublishOpen.value = false;
+  publish();
+}
+
+function publishCancel() {
+  confirmPublishOpen.value = false;
+}
 </script>
 
 <template>
-  <v-tooltip location="top">
-    <template #activator="{ props: actv }">
+  <v-tooltip
+    location="top"
+  >
+    <template
+      #activator="{ props: actv }"
+    >
       <v-btn
         v-bind="actv"
         variant="text"
@@ -53,7 +69,7 @@ async function onClick() {
         icon
         :loading="loading"
         :aria-label="btnLabel"
-        @click.stop="onClick"
+        @click="isDraft ? confirmPublishOpen = true : publish()"
       >
         <v-icon
           size="20"
@@ -63,4 +79,10 @@ async function onClick() {
     </template>
     <span>{{ btnLabel }}</span>
   </v-tooltip>
+
+  <ConfirmPublishPopup
+    :dialog-open="confirmPublishOpen"
+    :ok-clicked="publishOk"
+    :cancel-clicked="publishCancel"
+  />
 </template>
